+++ CHANGELOG.md
@@ -1,4 +1,10 @@
-# 0.3.0
+# 0.4.0
+
+### Refactor
++ Refactor custom_account view (Change fields to render)
++ Add accounts into envelope
+
+## 0.3.0
 + Small refactor on ping
 + Add setup.cfg for flake
 + Crypt passwords